Fini leaderboard

+++ Assets/Scripts/ScoreManager.cs
@@ -6,16 +6,17 @@ using UnityEngine.Events;
 
 public class ScoreManager : MonoBehaviour
 {
+
     [SerializeField]
     private TextMeshProUGUI inputScore;
 
     [SerializeField]
     private TMP_InputField inputName;
 
-    public UnityEvent<string, float> submitScoreEvent;
+    public UnityEvent<string, int> submitScoreEvent;
 
     public void SubmitScore()
     {
-        submitScoreEvent.Invoke(inputName.text, float.Parse(inputScore.text));
+        submitScoreEvent.Invoke(inputName.text, int.Parse(inputScore.text));
     }
 }
